refactor: 패키지 이동, 클래스명 변경에 따른 View, Controller 리팩토링
